@@ -6,31 +6,31 @@
   "author": "dry-tortuga",
   "license": "MIT",
   "dependencies": {
-    "@headlessui/react": "^1.4.3",
-    "@heroicons/react": "^1.0.5",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@web3-react/coinbase-wallet": "^8.2.3",
-    "@web3-react/core": "^8.2.3",
-    "@web3-react/metamask": "^8.2.4",
-    "@web3-react/walletconnect-v2": "^8.5.1",
+    "@tanstack/react-query": "^5.37.1",
+    "@wagmi/core": "^2.10.3",
     "bootstrap": "^5.1.3",
     "bootswatch": "^5.1.3",
     "buffer": "^6.0.3",
     "dayjs": "^1.11.9",
-    "ethers": "^5.7.2",
-    "react": "^17.0.2",
+    "ethers": "^6.12.1",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.4.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.7.2",
+    "viem": "^2.12.0",
+    "wagmi": "^2.9.6",
     "web3": "^1.7.1"
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "env-cmd": "^10.1.0"
+    "@eslint/compat": "^1.0.1",
+    "@eslint/js": "^9.3.0",
+    "env-cmd": "^10.1.0",
+    "eslint": "^9.3.0",
+    "eslint-plugin-react": "^7.34.2",
+    "globals": "^15.3.0"
   },
   "scripts": {
     "copy-abi-backend": "cp ../songbirdz-collection-backend/artifacts/contracts/SongBirdz.sol/SongBirdz.json ./src/abi/SongBirdz.json",
@@ -41,12 +41,6 @@
     "build:production": "env-cmd -f .env.production npm run build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
   },
   "browserslist": {
     "production": [
