@@ -1,22 +1,30 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Web3ReactProvider } from '@web3-react/core'
-import { Buffer } from 'buffer';
+import React from "react";
+import { createRoot } from "react-dom/client";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { WagmiProvider } from "wagmi";
+import { Buffer } from "buffer";
 
-import App from './App';
-import { WALLET_CONNECTORS } from './contexts/wallet';
+import App from "./App";
+import config from "./config";
 
-import 'bootswatch/dist/sketchy/bootstrap.min.css';
-import './index.css';
+import "bootswatch/dist/sketchy/bootstrap.min.css";
+import "./index.css";
 
-// Polyfill window buffer object for wallet connection logic
+const queryClient = new QueryClient();
+
+// Polyfill window buffer object for wallet connection logic (TODO)
 window.Buffer = window.Buffer || Buffer;
 
-ReactDOM.render(
+const container = document.getElementById("root");
+
+const root = createRoot(container);
+
+root.render(
 	<React.StrictMode>
-		<Web3ReactProvider connectors={WALLET_CONNECTORS}>
-			<App />
-		</Web3ReactProvider>
+		 <WagmiProvider config={config}>
+		 	<QueryClientProvider client={queryClient}>
+				<App />
+			</QueryClientProvider>
+		</WagmiProvider>
 	</React.StrictMode>,
-	document.getElementById('root')
 );
