@@ -1,34 +1,18 @@
-import { COLLECTION_BIRD_SIZE } from '../constants';
+import { COLLECTION_BIRD_SIZE } from "../constants";
 
-async function fetchBird(contract, id) {
+async function fetchBird(id, owner) {
 
-	if (!contract) {
-		return null;
-	}
-
-	console.debug('fetching bird #' + id);
+	console.debug("fetching bird #" + id);
 
 	const data = {
 		id,
 		name: `Songbird #${id}`,
-		owner: null,
+		owner,
 		species: null,
 		image: `${process.env.PUBLIC_URL}/images/${id}.jpg`,
 		imageLg: `${process.env.PUBLIC_URL}/images/${id}-lg.jpg`,
 		collection: Math.floor(id / COLLECTION_BIRD_SIZE),
 	};
-
-	// Fetch the owner data from the solidity contract
-	try {
-
-		data.owner = await contract.ownerOf(id);
-
-	} catch (error) {
-
-		// Does not have owner yet
-		console.debug(error);
-
-	}
 
 	// Fetch the off-chain metadata from the back-end server
 	const finalData = await populateMetadata(data);
@@ -56,8 +40,6 @@ async function populateMetadata(data) {
 				const responseData = await response.json();
 
 				finalData.species = responseData.species;
-				// finalData.image = responseData.image_onchain;
-				// finalData.imageLg = responseData.image;
 
 			}
 
@@ -68,16 +50,12 @@ async function populateMetadata(data) {
 			console.debug("--------------------------------------");
 
 			finalData.species = "UNIDENTIFIED";
-			// finalData.image = `${process.env.PUBLIC_URL}/images/unidentified.jpg`;
-			// finalData.imageLg = `${process.env.PUBLIC_URL}/images/unidentified.jpg`;
 
 		}
 
 	} else {
 
 		finalData.species = "UNIDENTIFIED";
-		// finalData.image = `${process.env.PUBLIC_URL}/images/unidentified.jpg`;
-		// finalData.imageLg = `${process.env.PUBLIC_URL}/images/unidentified.jpg`;
 
 	}
 
